refactored the code to fix the mistakes from the previous commit
