+++ src/components/pages/home/fetch-home-data.ts
@@ -4,6 +4,7 @@ import {
   fromPrismaTransaction,
   fromPrismaTransactionEnqueued,
 } from "@/lib/types";
+import { subDays, format } from "date-fns";
 import { l1PublicClient, l2PublicClient } from "@/lib/chains";
 import {
   fetchTokensPrices,
@@ -11,6 +12,33 @@ import {
   fetchLatestTransactions,
   fetchLatestTransactionsEnqueued,
 } from "@/lib/fetch-data";
+
+const fetchTransactionHistory = async () => {
+  const dates = Array.from({ length: 15 }, (_, i) => {
+    const date = subDays(new Date(), 14 - i);
+    return {
+      name: format(date, "MMM d"),
+      formattedDate: format(date, "yyyy-MM-dd"),
+    };
+  });
+
+  const responses = await Promise.all(
+    dates.map(async (dateObj) => {
+      const response = await fetch(
+        `https://api.coinbase.com/v2/prices/OP-USD/spot?date=${dateObj.formattedDate}`,
+      );
+      const res = await response.json();
+      return {
+        name: dateObj.name,
+        date: dateObj.formattedDate,
+        price: parseFloat(res.data.amount),
+        transactionsCount: Math.floor(Math.random() * 200) + 400,
+      };
+    }),
+  );
+
+  return responses;
+};
 
 const fetchHomeDataFromJsonRpc = async () => {
   const [latestL1BlockNumber, latestL2BlockNumber] = await Promise.all([
@@ -22,6 +50,7 @@ const fetchHomeDataFromJsonRpc = async () => {
     latestBlocks,
     { transactions: latestTransactions },
     { transactionsEnqueued: latestTransactionsEnqueued },
+    transactionHistory,
   ] = await Promise.all([
     fetchTokensPrices(),
     fetchLatestBlocks(latestL2BlockNumber),
@@ -31,12 +60,14 @@ const fetchHomeDataFromJsonRpc = async () => {
       "0x",
       latestL1BlockNumber,
     ),
+    fetchTransactionHistory(),
   ]);
   return {
     tokensPrices,
     latestBlocks: latestBlocks.slice(0, 6),
     latestTransactions: latestTransactions.slice(0, 6),
     latestTransactionsEnqueued: latestTransactionsEnqueued.slice(0, 10),
+    transactionHistory,
   };
 };
 
@@ -46,6 +77,7 @@ const fetchHomeDataFromDatabase = async () => {
     latestBlocks,
     latestTransactions,
     latestTransactionsEnqueued,
+    transactionHistory,
   ] = await Promise.all([
     fetchTokensPrices(),
     prisma.block.findMany({
@@ -61,6 +93,7 @@ const fetchHomeDataFromDatabase = async () => {
       orderBy: { timestamp: "desc" },
       take: 10,
     }),
+    fetchTransactionHistory(),
   ]);
   return {
     tokensPrices,
@@ -73,9 +106,10 @@ const fetchHomeDataFromDatabase = async () => {
     latestTransactionsEnqueued: latestTransactionsEnqueued.map(
       fromPrismaTransactionEnqueued,
     ),
+    transactionHistory,
   };
 };
 
-export const fetchHomeData = process.env.DATABASE_URL
+export  const fetchHomeData = process.env.DATABASE_URL
   ? fetchHomeDataFromDatabase
   : fetchHomeDataFromJsonRpc;
